+++ server/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - PyPika: Athena translator now uses `DATE_PARSE` rather than `TIMESTAMP` for the `todate` step
+- PyPika: Postgres and Redshift translator now use `TO_TIMESTAMP` rather than `TIMESTAMP` for the `todate` step
 
 ## [0.25.4] - 2022-10-06